+++ OUDS/Core/Themes/Orange/Tests/Values/ComponentTokens/ThemeOverrideOfIconComponentTokensTests.swift
@@ -37,8 +37,13 @@ struct ThemeOverrideOfIconComponentTokensTests {
         #expect(inheritedTheme.icon.colorContentStatusWarningInternalShape == MockThemeIconComponentTokenProvider.mockThemeIconColor)
     }
 
-    @Test func inheritedThemeCanOverrideIconComponentTokenColorContentDefault() throws {
-        #expect(inheritedTheme.icon.colorContentDefault != abstractTheme.icon.colorContentDefault)
-        #expect(inheritedTheme.icon.colorContentDefault == MockThemeIconComponentTokenProvider.mockThemeIconColor)
+    @Test func inheritedThemeCanOverrideIconComponentTokenIconColorContentStatusWarningInverseExternalShape() throws {
+        #expect(inheritedTheme.icon.colorContentStatusWarningInverseExternalShape != abstractTheme.icon.colorContentStatusWarningInverseExternalShape)
+        #expect(inheritedTheme.icon.colorContentStatusWarningInverseExternalShape == MockThemeIconComponentTokenProvider.mockThemeIconColor)
+    }
+
+    @Test func inheritedThemeCanOverrideIconComponentTokenIconColorContentStatusWarningInverseInternalShape() throws {
+        #expect(inheritedTheme.icon.colorContentStatusWarningInverseInternalShape != abstractTheme.icon.colorContentStatusWarningInverseInternalShape)
+        #expect(inheritedTheme.icon.colorContentStatusWarningInverseInternalShape == MockThemeIconComponentTokenProvider.mockThemeIconColor)
     }
 }