+++ OUDS/Core/Themes/Wireframe/Sources/WireframeTheme.swift
@@ -49,9 +49,9 @@ import SwiftUI
 ///
 /// ## Tokens versions
 ///
-/// - Core OUDS version: 1.9.0
-/// - Core Wireframe version: 1.3.0
-/// - Brand Wireframe version: 2.3.0
+/// - Core OUDS version: 1.10.0
+/// - Core Wireframe version: 1.4.0
+/// - Brand Wireframe version: 2.4.0
 ///
 /// - Since: 0.17.0
 public final class WireframeTheme: OUDSTheme, @unchecked Sendable {