+++ OUDS/Core/Themes/Wireframe/Tests/Values/SemanticTokens/WireframeTheme+BorderSemanticTokensTests.swift
@@ -19,7 +19,7 @@ import Testing
 /// Indeed, each future generation of Swift code may break theses tests because there are new values.
 /// In the semantics of **border semantic tokens**, there will be some unchanged things like relationships between tokens.
 /// Thus this tests class just checks if such relationships are still here whatever the values at the end.
-struct SoshThemeBorderSemanticTokensTests {
+struct WireframeThemeBorderSemanticTokensTests {
 
     private var theme: OUDSTheme
 
@@ -28,9 +28,9 @@ struct SoshThemeBorderSemanticTokensTests {
     }
 
     @Test func widthSemanticTokensAreBiggerAndBigger() throws {
-        #expect(theme.borders.widthNone < theme.borders.widthDefault)
-        #expect(theme.borders.widthDefault <= theme.borders.widthThin)
-        #expect(theme.borders.widthThin < theme.borders.widthMedium)
+        #expect(theme.borders.widthNone < theme.borders.widthThin)
+        #expect(theme.borders.widthThin < theme.borders.widthDefault)
+        #expect(theme.borders.widthDefault <= theme.borders.widthMedium)
         #expect(theme.borders.widthMedium < theme.borders.widthThick)
         #expect(theme.borders.widthThick < theme.borders.widthThicker)